vertices: read the dtype the store declares, not an assumed float32
Answers the review question on the vlen layout: "where is the metadata
about how to decode the variable_length_bytes into an array?"

It was always stored -- create_vertices_array stamps a `dtype` attribute
-- but the reader ignored it. read_chunk_vertices defaulted to
`dtype=np.float32`, and coarsen, rebin, lazy/level, and both validators
hardcoded it on top. So the spec's "readers MUST honour the stored dtype"
was a rule the library itself broke.

The default now reads the attribute (new `vertices_dtype()` helper); pass
`dtype=` only to override. On a float64 store:

    before (default float32):  [[0.0, 1.875, 0.0], [2.0, 0.0, 2.125]]
    after  (declared float64): [[1.0, 2.0, 3.0], [4.0, 5.0, 6.0], ...]

Note it did not raise -- a cell is a flat buffer with no inline header, so
an assumed dtype cannot be checked against anything. It just returns
different numbers, at twice the row count, looking entirely plausible.
That is the point: the container type (`variable_length_bytes`) says
nothing about the payload, so the only correct source is the attribute the
writer stamped. Guessing is not a fallback here, it is a silent corruption.

This is the same regression the vlen migration introduced in the docs:
under Option G, Zarr's own `data_type: float32` and `shape [...,N_max,3]`
were self-describing and a reader needed no ZVF metadata at all. Moving to
vlen shifted dtype, D and N onto ZVF attributes plus derivation. The spec
page picked that up; the reader had not.

+++ zarr_vectors/lazy/level.py
@@ -225,7 +225,7 @@ class ZVLevel:
                 continue
             try:
                 groups.extend(
-                    read_chunk_vertices(self._group, cc, dtype=np.float32, ndim=ndim)
+                    read_chunk_vertices(self._group, cc, ndim=ndim)
                 )
             except Exception:
                 continue
